♻️ Move getCountryCurrency into the billing package
It is a pure mapping with no app dependency, and the landing page will
need it for currency detection too, so it lives beside the displayed
currency allowlist instead of in the web app.
